refactor: change variant -> option in names
