+++ src/pages/Configurators/components/UI/OptionItems/ExteriorOptionsItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExteriorOptionsItem'
