+++ src/pages/Configurators/components/UI/OptionItems/InteriorOptionsItem/InteriorOptionsItem.scss
