+++ src/pages/Configurators/components/UI/OptionItems/InteriorOptionsItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './InteriorOptionsItem'
