+++ src/pages/Configurators/components/UI/OptionsList/OptionsList.module.scss
