+++ src/pages/Configurators/components/UI/OptionsList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OptionsList'
