--- src/pages/Configurators/components/UI/VariantsItems/PrimaryVariantsItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PrimaryVariantsItem'
