--- src/pages/Configurators/components/UI/VariantsItems/SecondaryVariantsItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SecondaryVariantsItem'
