--- src/pages/Configurators/components/UI/VariantsItems/TertiaryVariantsItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TertiaryVariantsItem'
