--- src/pages/Configurators/components/UI/VariantsList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './VariantsList'
